Separated Nav logic out of render method
